Reduce MAX-Y and acceleration after heatbed testing
The y-axis speed worked really well and smoothly with a fast
acceleration and at high speeds (18000 steps/sec) when it was
unloaded.  But tests with the ~1100 gram heat bed attached
show that the maximum attainable speed dropped about 20%.  I
can safely get up to about 14000 now.  But I set the max to 12000
to keep a bit of safety margin.

Similar tests with the Z-axis (no loading) so I've bumped the
speed there up a bit, too.