Add testcases for velocity attained
WARNING this script will overwrite your config.h

The velocity-test.sh script will compile the simulator with a set
of canned config parameters and for a range of different ACCELERATION
values.  Then it runs the simulator against a range of different
target speeds, with and without lookahead enabled.  Finally it plots
the results in gnuplot to the screen.

The actual speed is calculated from the trace file using an awk
script to calculate the differential of the position w.r.t time.
Time is on the X axis and is given in minutes.  Speed on the Y
axis is given in mm/min (same as feedrate).