More: Flesh out exponential planner research
exponential.c will produce two plots of velocity and position which
appear to be accurate.  Plot it on gnuplot like this:

  plot "exponential.out" using 2:3 with lines, '' u 2:4 with lines

Now suports piece-wise integration of the velocity profile to make up
the position calculation.

Issue 1:
When we cannot reach vmax, we do not try hard enough.  The move gets
"blended" by rampup/rampdown overlap severely.  This is not a problem
for our 'exponential' linear approximation plans, but it does get in the
way of testing this implementation.

Issue 2:
We use floating point everywhere here.  We need to use integers if we
want to port this into Teacup.