Flesh exponential planner research
exponential.c will produce two plots of velocity and position which
appear to be accurate.  Plot it on gnuplot like this:

  plot "exponential.out" using 2:3 with lines, '' u 2:4 with lines

Issue 1:
The position plot is based on separate acceleration calculations
from the velocity one.  We want instead to have a piece-wise integration
of the velocity profile to make up the position calculation.

Issue 2:
We use floating point everywhere here.  We need to use integers if we
want to port this into Teacup.