More: Flesh out exponential planner research
exponential.c will produce two plots of velocity and position which
appear to be accurate.  Plot it on gnuplot like this:

  plot "exponential.out" using 2:3 with lines, '' u 2:4 with lines

Now suports piece-wise integration of the velocity profile to make up
the position calculation.

When we cannot reach vmax, we adjust vmax so it is reachable without any
overlap.  This may not be needed for 'exponential' linear approximations.

TBD:
We use floating point everywhere here.  We need to use integers if we
want to port this into Teacup.