WIP: Use uint64_t integer math
We need to get down to 8-bit-efficient math without any FP math.
We lose too much precision with floats, so we have been using
doubles.  Let's start reigning that back in.

Convert the doubles back to floats.  To prevent the resulting
accumulated error, start using all integer math for step planning
during our piece-wise integration.  Calculate the overflow in a
remainder variable, and determine when to step the quotient by
comparing it to the divisor.

We will use 64-bit ints for now, and try to simplify that again
in a future commit.