Try to predict the next step in real-time
This fails because we always want to be called before our next step
so we can continue to integrate before we miss a step.  This is
fairly reliable for most reasonable values of A, V and DX, but it
requires far more interrupts than are truly necessary.  To fix that,
we will turn to the clock-based math method which is already used in
Teacup.