Corrected v slope and dsteps slope
These work and for these limited test cases, the number of steps
works out, too.  But step predictions are poor and the real-time
integration is frenetic for now.  There may be cases where a late
dstep causes two quick steps in succession, still.  I'm trying to
work those out in the next commit.