Remove vestiges of "genius" idea, not currently used.
Seems like this still has merit.  For example, if we find that
remainder is huge after a step, we must not be stepping fast
enough so we should bias our next step forward a bit.  But if
we do that we risk compensating for error all at once and causing
a missed step.  Best not to risk that unless we can always manage
to keep remainder very low by fine-tuning dStep very well.