Track 16-bit timer values
Move to 16-bit quanta instead of 64-bit time-of-day trackers.  This will
make the atmel AVR happier and faster, and it also moves in the direction
of supporting a linear velocity profile array in the future.

Although, it is beginning to look like we might need to store acceleration
values instead of velocity ones, that will actually make it easier for us
to calculate the velocity curve over time.