Convert to acceleration-profile
I've been calling this a velocity profile all this time, but we
might as well use acceleration instead.  The only velocity we really
work with is our accumulated velocity and vmax.  All the rest of the
math works with acceleration.

Switching this way has a pleasing effect on the code.  It causes
us not to subtract v(Td-Ts) from our current velocity but instead to
add the negative acceleration instead.  This also fits better with
our phase approach we will need for the later acceleration profile.
I think.