BROKEN: Add parser contexts to isolate various cmd sources
Proof-Of-Concept code

Commands may come from UART, SD Card, or canned commands.  Keep the
different streams separate to avoid toe-stepping.

This code is broken because different streams now have their own
individual ideas about the last "state" of the machine. Only the
last command processed is correct, though, and the other command
states need to be updated when a new command is processed.  But their
state may be mid-parse, so this would break their expectations.

This can probably be fixed by splitting out the "machine state" from
the command-parser, but it will affect things like feedrates, and
relative positions in a complex way.