Reduce lookahead "error" to "warning"
The marlin firmware reportedly reports "Error" instead of "!!", indicating
a machine failure which is followed by a full power-down.  The Octoprint
GCode-sender assumes a reported Error means the print has failed and
the machine turned off.

In Teacup we report an "Error" when lookahead was too slow to join
movements, but this is interpreted wrongly by Octoprint who then stops
the job and leaves the printer idle with all the heaters running.

Change this "Error" to a "Warning" to avoid this problem.  I have not
tested the "Warning" report in Octoprint yet to ensure it doesn't cause
some other calamity, but I will soon.