simulator: support endstop simulation
Provide a simulated, simplified representation of the action of
mechanical endstop switches which turn on and off at slightly different
amounts of pressure. When any axis moves past -10, simulate endstop
"on" condition. When the axis moves to 0, simulate the endstop "off"
condition.

This support allows the simulation of G28 (home) commands.