dda: record position when endstop is triggered
When we hit an endstop we fudge the numbers in move_state to pretend we
reached the end of our movement. This is ok because home is about to
zero our position anyway.  But we may want to probe for an endstop and
return the position where it was found. We can't do that with the current
destructive position movement.

When we reach an endstop and offset our position to fudge the stop,
remember how much we fudged it in endstop_skipped_steps.  When we do
finish stopping, fix up the move_state counters to look as if we did
make those steps, and calculate our new location in endstop_position_um.